ops/requirements.txt
boto3
psycopg2-binary
torch
numpy 